+++ composeApp/src/commonMain/kotlin/com/github/freshmorsikov/moviematcher/util/NumberUtil.kt
@@ -10,7 +10,7 @@ fun Int.pow(x: Int): Int {
     return this.toDouble().pow(x).toInt()
 }
 
-fun Int.toAmountFormat(): String {
+fun Long.toAmountFormat(): String {
     return this.toString()
         .reversed()
         .windowed(size = 3, step = 3, partialWindows = true) {
